End of Day Push 27/08/22
FEATURE: Added a sign in Level 1 to teach the player the controls.
FEATURE: Added a Save Game Blueprint (Not yet implemented)

Spent a lot of the day trying to figure out the UI and why Volume sliders are not working correctly, the same with the check box for Fullscreen. The Resolution picker works, but it seems to fight with the Fullscreen check box and will go back and forth between windowed and not windowed whenever the pause menu is opened.